backend/requirements.txt
fastapi
uvicorn
pymongo==4.6.0
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart